Revert "lib: cheri: Support smode software running in integer pointer mode"
This change originally add an option to allow the opensbi ecall to form
a capability pointer from hsm ecall and legacy ecall if the pointer is
invalid and assume it calling from smode caller running in integer pointer
execution mode if CHERI hybrid mode is supported.

The caller execution mode can be detected by the mode bit of mepc. To
support hybrid smode execution. A valid capability pointer is only formed
by an invalid pointer passed from a smode caller running in integer pointer
exection mode to keep the opensbi running in capability pointer mode to
support a smode executable.