include: sbi: cheri: Add cheri_is_integer_pointer_mode_execution()
cheri_is_integer_pointer_mode_execution() is added to pass an excutable
capability pointer to check if it is in integer pointer mode when
__riscv_zcherihybrid is supported. It returns true (1) if the mode bit in
exe_cap is 1 (integer pointer execution mode) or false (0) if the mode bit
in exe_cap is 0 (capability pointer execution mode).

It will always return false (0) if on a CHERI RISCV purecap core
(__riscv_zcheripurecap) when __riscv_zcherihybrid is not supported.